Moved WeaponType to static classes to test save compatibility
