+++ src/com/base/game/inventory/clothing/ClothingType.java
@@ -508,8 +508,7 @@ public class ClothingType {
 			null,
 			Colour.allClothingColours) {
 		
-		
-				private static final long serialVersionUID = 1L;
+		private static final long serialVersionUID = 1L;
 
 		@Override
 		public String equipText(GameCharacter clothingOwner, GameCharacter clothingRemover, boolean rough, AbstractClothing clothing, boolean applyEffects) {
@@ -549,8 +548,7 @@ public class ClothingType {
 			null,
 			Colour.allClothingColours) {
 		
-		
-				private static final long serialVersionUID = 1L;
+		private static final long serialVersionUID = 1L;
 
 		@Override
 		public String equipText(GameCharacter clothingOwner, GameCharacter clothingRemover, boolean rough, AbstractClothing clothing, boolean applyEffects) {
