+++ src/com/base/game/inventory/weapon/WeaponType.java
@@ -1,11 +1,8 @@
 package com.base.game.inventory.weapon;
 
-import java.io.IOException;
-import java.io.InputStream;
+import java.lang.reflect.Field;
 import java.util.ArrayList;
-import java.util.EnumMap;
 import java.util.List;
-import java.util.Map;
 
 import com.base.game.character.GameCharacter;
 import com.base.game.character.attributes.Attribute;
@@ -13,25 +10,20 @@ import com.base.game.combat.DamageLevel;
 import com.base.game.combat.DamageType;
 import com.base.game.combat.DamageVariance;
 import com.base.game.combat.Spell;
-import com.base.game.dialogue.utils.UtilText;
 import com.base.game.inventory.InventorySlot;
 import com.base.game.inventory.Rarity;
-import com.base.main.Main;
-import com.base.utils.Colour;
 import com.base.utils.Util;
 import com.base.utils.Util.ListValue;
 import com.base.utils.Util.Value;
 
 /**
- * @since 0.1.0
- * @version 0.1.75
+ * @since 0.1.84
+ * @version 0.1.84
  * @author Innoxia
  */
-public enum WeaponType {
-
-	// MAGIC CRYSTALS:
-	// MELEE
-	MELEE_CHAOS_RARE("an",
+public class WeaponType {
+	
+	public static AbstractWeaponType MELEE_CHAOS_RARE = new AbstractWeaponType("an",
 			"it",
 			"opaque demonstone",
 			"A common type of demonstone, the power of which can be harnessed as a weapon."
@@ -44,7 +36,9 @@ public enum WeaponType {
 			DamageVariance.MEDIUM,
 			null,
 			null) {
-
+		
+		private static final long serialVersionUID = 1L;
+				
 		@Override
 		public String equipText(GameCharacter character) {
 			return "You focus on the energy in the crystal. As you do so, it dissolves and flows into your body, granting you the ability to perform magical attacks.";
@@ -59,8 +53,9 @@ public enum WeaponType {
 		public String getAttackDescription(GameCharacter character, GameCharacter target) {
 			return genericMeleeAttackDescription(character, target);
 		}
-	},
-	MELEE_CHAOS_EPIC("a",
+	};
+	
+	public static AbstractWeaponType MELEE_CHAOS_EPIC = new AbstractWeaponType("a",
 			"it",
 			"misty demonstone",
 			"A powerful demonstone, the power of which can be harnessed as a weapon."
@@ -74,6 +69,8 @@ public enum WeaponType {
 			null,
 			null) {
 
+		private static final long serialVersionUID = 1L;
+
 		@Override
 		public String equipText(GameCharacter character) {
 			return "You focus on the energy in the crystal. As you do so, it dissolves and flows into your body, granting you the ability to perform magical attacks.";
@@ -88,8 +85,9 @@ public enum WeaponType {
 		public String getAttackDescription(GameCharacter character, GameCharacter target) {
 			return genericMeleeAttackDescription(character, target);
 		}
-	},
-	MELEE_CHAOS_LEGENDARY("a",
+	};
+	
+	public static AbstractWeaponType MELEE_CHAOS_LEGENDARY = new AbstractWeaponType("a",
 			"it",
 			"clear demonstone",
 			"An extremely powerful demonstone, the power of which can be harnessed as a weapon."
@@ -103,6 +101,8 @@ public enum WeaponType {
 			Util.newHashMapOfValues(new Value<Attribute, Integer>(Attribute.STRENGTH, 5)),
 			Util.newArrayListOfValues(new ListValue<Spell>(Spell.FIREBALL_1), new ListValue<Spell>(Spell.FIREBALL_1), new ListValue<Spell>(Spell.FIREBALL_1))) {
 
+		private static final long serialVersionUID = 1L;
+
 		@Override
 		public String equipText(GameCharacter character) {
 			return "You focus on the energy in the crystal. As you do so, it dissolves and flows into your body, granting you the ability to perform magical attacks.";
@@ -117,10 +117,10 @@ public enum WeaponType {
 		public String getAttackDescription(GameCharacter character, GameCharacter target) {
 			return genericMeleeAttackDescription(character, target);
 		}
-	},
+	};
 
 	// OFFHAND
-	OFFHAND_CHAOS_RARE("a",
+	public static AbstractWeaponType OFFHAND_CHAOS_RARE = new AbstractWeaponType("a",
 			"it",
 			"chaos feather",
 			"A magical feather, the power of which can be harnessed as a weapon."
@@ -133,6 +133,8 @@ public enum WeaponType {
 			DamageVariance.HIGH,
 			null,
 			null) {
+
+		private static final long serialVersionUID = 1L;
 
 		@Override
 		public String equipText(GameCharacter character) {
@@ -148,8 +150,9 @@ public enum WeaponType {
 		public String getAttackDescription(GameCharacter character, GameCharacter target) {
 			return genericRangedAttackDescription(character, target);
 		}
-	},
-	OFFHAND_CHAOS_EPIC("a",
+	};
+	
+	public static AbstractWeaponType OFFHAND_CHAOS_EPIC = new AbstractWeaponType("a",
 			"it",
 			"chaos feather",
 			"A well-preserved magical feather, the power of which can be harnessed as a weapon."
@@ -162,6 +165,9 @@ public enum WeaponType {
 			DamageVariance.HIGH,
 			null,
 			null) {
+		
+		private static final long serialVersionUID = 1L;
+
 		@Override
 		public String equipText(GameCharacter character) {
 			return "You focus on the energy in the feather. As you do so, it dissolves and flows into your body, granting you the ability to perform magical attacks at range.";
@@ -178,202 +184,30 @@ public enum WeaponType {
 		}
 	};
 
-	private static List<WeaponType> rareWeapons;
+	public static List<AbstractWeaponType> rareWeapons = new ArrayList<>(), allweapons = new ArrayList<>();
 
 	static {
-		rareWeapons = new ArrayList<>();
-		for (WeaponType w : WeaponType.values())
-			if (w.getRarity() == Rarity.RARE)
-				rareWeapons.add(w);
-	}
-
-	private String determiner, pronoun, name, description, pathName;
-	protected DamageLevel damageLevel;
-	protected DamageVariance damageVariance;
-	private InventorySlot slot;
-	private List<DamageType> availableDamageTypes;
-	private Rarity rarity;
-	private Map<Attribute, Integer> attributeModifiers;
-	private Map<DamageType, String> SVGStringMap;
-	private List<Spell> spells;
-
-	private WeaponType(String determiner, String pronoun, String name, String description, InventorySlot slot, String pathName, Rarity rarity, List<DamageType> availableDamageTypes, DamageLevel damageLevel, DamageVariance damageVariance,
-			Map<Attribute, Integer> attributeModifiers, List<Spell> spells) {
-
-		this.determiner = determiner;
-		this.pronoun = pronoun;
-		this.name = name;
-		this.description = description;
-		this.rarity = rarity;
-
-		this.slot = slot;
-		this.availableDamageTypes = availableDamageTypes;
-
-		this.damageLevel = damageLevel;
-		this.damageVariance = damageVariance;
-
-		this.pathName = pathName;
-		this.attributeModifiers = attributeModifiers;
-		this.spells = spells;
-
-		SVGStringMap = new EnumMap<>(DamageType.class);
-		for (DamageType dt : this.availableDamageTypes)
-			try {
-				InputStream is = this.getClass().getResourceAsStream("/com/base/res/weapons/"
-						+ pathName
-						+ ".svg");
-				String s = Util.inputStreamToString(is);
-
-				s = s.replaceAll("#ff2a2a", dt.getMultiplierAttribute().getColour().getShades()[0]);
-				s = s.replaceAll("#ff5555", dt.getMultiplierAttribute().getColour().getShades()[1]);
-				s = s.replaceAll("#ff8080", dt.getMultiplierAttribute().getColour().getShades()[2]);
-				s = s.replaceAll("#ffaaaa", dt.getMultiplierAttribute().getColour().getShades()[3]);
-				s = s.replaceAll("#ffd5d5", dt.getMultiplierAttribute().getColour().getShades()[4]);
-				SVGStringMap.put(dt, s);
-
-				is.close();
-			} catch (IOException e) {
-				e.printStackTrace();
-			}
-
-	}
-
-	public static AbstractWeapon generateWeapon(WeaponType wt, DamageType dt) {
-		DamageType damageType = dt;
-
-		if (wt.getAvailableDamageTypes() != null)
-			if (!wt.getAvailableDamageTypes().contains(dt))
-				dt = wt.getAvailableDamageTypes().get(Util.random.nextInt(wt.getAvailableDamageTypes().size()));
-
-		return new AbstractWeapon(wt, damageType) {
-			private static final long serialVersionUID = 1L;
-
-			@Override
-			public String onEquip(GameCharacter character) {
-				if (character.isPlayer()) {
-					if (Main.game.getPlayer().getWeaponsDiscovered().add(wt)) {
-						Main.game.getPlayer().setNewWeaponDiscovered(true);
-						Main.game.getTextEndStringBuilder().append(
-								"<p style='text-align:center;'>"
-									+ "<b style='color:"+Colour.GENERIC_EXCELLENT.toWebHexString()+";'>New entry in your phone's encyclopedia:</b>"
-									+ "</br>"
-									+ "<b>Weapon:</b> <b style='color:"+wt.getRarity().getColour().toWebHexString()+";'>"+Util.capitaliseSentence(wt.getName())+"</b>"
-								+ "</p>");
+		
+		Field[] fields = WeaponType.class.getFields();
+		
+		for(Field f : fields){
+			
+			if (AbstractWeaponType.class.isAssignableFrom(f.getType())) {
+				
+				AbstractWeaponType weapon;
+				try {
+					weapon = ((AbstractWeaponType) f.get(null));
+					
+					allweapons.add(weapon);
+					
+					if (weapon.getRarity() == Rarity.RARE) {
+						rareWeapons.add(weapon);
 					}
+					
+				} catch (IllegalArgumentException | IllegalAccessException e) {
+					e.printStackTrace();
 				}
-				return wt.equipText(character);
 			}
-
-			@Override
-			public String onUnequip(GameCharacter character) {
-				return wt.unequipText(character);
-			}
-		};
-	}
-
-	/**
-	 * Generates a weapon with random damage type
-	 * 
-	 * @param wt
-	 * @param level
-	 * @return
-	 */
-	public static AbstractWeapon generateWeapon(WeaponType wt) {
-		return WeaponType.generateWeapon(wt, wt.getAvailableDamageTypes().get(Util.random.nextInt(wt.getAvailableDamageTypes().size())));
-	}
-
-	public abstract String equipText(GameCharacter character);
-
-	public abstract String unequipText(GameCharacter character);
-	
-	public abstract String getAttackDescription(GameCharacter character, GameCharacter target);
-
-	
-	public static String genericMeleeAttackDescription(GameCharacter character, GameCharacter target) {
-		if(character.isPlayer()) {
-			return UtilText.returnStringAtRandom(
-					"Darting forwards, you deliver a solid punch to [npc.name]'s [npc.arm].",
-					
-					"You throw a punch at [npc.name], grinning as you feel it connect with [npc.her] [npc.arm].",
-					
-					"You kick out at [npc.name], smiling to yourself as you feel your foot connect with [npc.her] [npc.leg].");
-			
-		} else {
-			return UtilText.returnStringAtRandom(
-					character.getName("The")+" punches you!");
-			
-		}	
-	}
-	
-	public static String genericRangedAttackDescription(GameCharacter character, GameCharacter target) {
-		if(character.isPlayer()) {
-			return UtilText.returnStringAtRandom(
-					"Punching your fist out towards [npc.name], a bolt of arcane energy shoots out to strike [npc.her] [npc.arm].",
-					
-					"Striking out towards [npc.name], a bolt of arcane energy shoots out of your fist to connect with [npc.her] [npc.arm].",
-					
-					"You kick out at [npc.name], and as you do, a bolt of arcane energy shoots out of your foot to connect with [npc.her] [npc.leg].");
-			
-		} else {
-			return UtilText.returnStringAtRandom(
-					character.getName("The")+" punches you!");
-			
-		}	
-	}
-	
-	public String getDeterminer() {
-		return determiner;
-	}
-
-	public String getPronoun() {
-		return pronoun;
-	}
-
-	public String getName() {
-		return name;
-	}
-
-	public String getDescription() {
-		return description;
-	}
-
-	public Rarity getRarity() {
-		return rarity;
-	}
-
-	public InventorySlot getSlot() {
-		return slot;
-	}
-
-	public String getPathName() {
-		return pathName;
-	}
-
-	public DamageLevel getDamageLevel() {
-		return damageLevel;
-	}
-
-	public DamageVariance getDamageVariance() {
-		return damageVariance;
-	}
-
-	public Map<Attribute, Integer> getAttributeModifiers() {
-		return attributeModifiers;
-	}
-
-	public List<DamageType> getAvailableDamageTypes() {
-		return availableDamageTypes;
-	}
-
-	public Map<DamageType, String> getSVGStringMap() {
-		return SVGStringMap;
-	}
-
-	public List<Spell> getSpells() {
-		return spells;
-	}
-
-	public static List<WeaponType> getRareWeapons() {
-		return rareWeapons;
+		}
 	}
 }
