+++ src/com/base/utils/WeaponRarityComparator.java
@@ -2,18 +2,18 @@ package com.base.utils;
 
 import java.util.Comparator;
 
-import com.base.game.inventory.weapon.WeaponType;
+import com.base.game.inventory.weapon.AbstractWeaponType;
 
 /**
  * Compares by rarity.
  * 
  * @since 0.1.2
- * @version 0.1.63
+ * @version 0.1.84
  * @author Innoxia
  */
-public class WeaponRarityComparator implements Comparator<WeaponType> {
+public class WeaponRarityComparator implements Comparator<AbstractWeaponType> {
 	@Override
-	public int compare(WeaponType first, WeaponType second) {
+	public int compare(AbstractWeaponType first, AbstractWeaponType second) {
 		int result = first.getRarity().compareTo(second.getRarity());
 		if (result != 0) {
 			return result;
